requirements.txt
azure-storage-blob
pyarrow
ultralytics
pyyaml
opencv-python-headless
numpy
pandas
imutils
Pillow
psutil